requirements.txt
numpy
confluent-kafka
tensorflow-gpu
